Update security.yml
Changing trufflehog to immutable version. Unauthorized manipulation of artifacts might be difficult to spot. For example, this may result in using images with malicious code. Also, intended major changes, which are automatically used in an image used might break the functionality. Pinning of artifacts ensure that changes are performed only when intended.